requirements.txt
pandas
numpy
mysql-connector-python
fastapi[all]
requests
datetime


